% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_obs.R
\name{count_obs}
\alias{count_obs}
\title{Count the Number of Observations of Categorical Variables}
\usage{
count_obs(data, group_vars, .groups = "drop")
}
\arguments{
\item{data}{A data frame;  this parameter was named \code{data} to make it obvious that this should be the
raw data frame from which the user wants to extract information.}

\item{group_vars}{The grouping variable(s). One or more grouping variables may be used. The grouping
variables specify which categorical variable(s) in \code{data} the function should count observations in.
The grouping variable must be of class chr, fct, or date. This parameter was named \code{group_vars} to
make it obvious that this variable should be used to specify the names of the variables that should
be used for grouping the data when the function is called.}
}
\value{
A tibble listing the number of observations of each value of a categorical variable or
combination of categorical variables specified in \code{group_vars} from the data frame \code{data}.
}
\description{
\code{count_obs} returns a summary tibble listing the number of observations for each value of a categorical
variable or combination of categorical variables in a specified data frame. The grouping variables must
belong to one of the following classes: chr, fct, or date.
}
\examples{
count_obs(df, categorical_variable_1)
count_obs(dataframe, c(var1, var2, var3))
}
